[fix](Nereids): split INNER and OUTER.
