+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/RuleSet.java
@@ -17,10 +17,12 @@
 
 package org.apache.doris.nereids.rules;
 
+import org.apache.doris.nereids.rules.exploration.join.InnerJoinLAsscom;
+import org.apache.doris.nereids.rules.exploration.join.InnerJoinLAsscomProject;
 import org.apache.doris.nereids.rules.exploration.join.JoinCommute;
 import org.apache.doris.nereids.rules.exploration.join.JoinCommuteProject;
-import org.apache.doris.nereids.rules.exploration.join.JoinLAsscom;
-import org.apache.doris.nereids.rules.exploration.join.JoinLAsscomProject;
+import org.apache.doris.nereids.rules.exploration.join.OuterJoinLAsscom;
+import org.apache.doris.nereids.rules.exploration.join.OuterJoinLAsscomProject;
 import org.apache.doris.nereids.rules.exploration.join.SemiJoinLogicalJoinTranspose;
 import org.apache.doris.nereids.rules.exploration.join.SemiJoinLogicalJoinTransposeProject;
 import org.apache.doris.nereids.rules.exploration.join.SemiJoinSemiJoinTranspose;
@@ -56,8 +58,10 @@ public class RuleSet {
     public static final List<Rule> EXPLORATION_RULES = planRuleFactories()
             .add(JoinCommute.LEFT_DEEP)
             .add(JoinCommuteProject.LEFT_DEEP)
-            .add(JoinLAsscom.INNER)
-            .add(JoinLAsscomProject.INNER)
+            .add(InnerJoinLAsscom.INSTANCE)
+            .add(InnerJoinLAsscomProject.INSTANCE)
+            .add(OuterJoinLAsscom.INSTANCE)
+            .add(OuterJoinLAsscomProject.INSTANCE)
             .add(SemiJoinLogicalJoinTranspose.LEFT_DEEP)
             .add(SemiJoinLogicalJoinTransposeProject.LEFT_DEEP)
             .add(SemiJoinSemiJoinTranspose.INSTANCE)
@@ -90,20 +94,20 @@ public class RuleSet {
             .build();
 
     public static final List<Rule> LEFT_DEEP_TREE_JOIN_REORDER = planRuleFactories()
-            .add(JoinCommute.LEFT_DEEP)
-            .add(JoinLAsscom.INNER)
-            .add(JoinLAsscomProject.INNER)
-            .add(JoinLAsscom.OUTER)
-            .add(JoinLAsscomProject.OUTER)
+            // .add(JoinCommute.LEFT_DEEP)
+            // .add(JoinLAsscom.INNER)
+            // .add(JoinLAsscomProject.INNER)
+            // .add(JoinLAsscom.OUTER)
+            // .add(JoinLAsscomProject.OUTER)
             // semi join Transpose ....
             .build();
 
     public static final List<Rule> ZIG_ZAG_TREE_JOIN_REORDER = planRuleFactories()
-            .add(JoinCommute.ZIG_ZAG)
-            .add(JoinLAsscom.INNER)
-            .add(JoinLAsscomProject.INNER)
-            .add(JoinLAsscom.OUTER)
-            .add(JoinLAsscomProject.OUTER)
+            // .add(JoinCommute.ZIG_ZAG)
+            // .add(JoinLAsscom.INNER)
+            // .add(JoinLAsscomProject.INNER)
+            // .add(JoinLAsscom.OUTER)
+            // .add(JoinLAsscomProject.OUTER)
             // semi join Transpose ....
             .build();
 
@@ -119,7 +123,7 @@ public class RuleSet {
             // .add(JoinExchangeLeftProject.INNER)
             // .add(JoinExchangeRightProject.INNER)
             // .add(JoinRightAssociate.OUTER)
-            .add(JoinLAsscom.OUTER)
+            // .add(JoinLAsscom.OUTER)
             // semi join Transpose ....
             .build();
