+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/RuleType.java
@@ -120,9 +120,10 @@ public enum RuleType {
     TEST_EXPLORATION(RuleTypeClass.EXPLORATION),
     LOGICAL_JOIN_COMMUTATE(RuleTypeClass.EXPLORATION),
     LOGICAL_LEFT_JOIN_ASSOCIATIVE(RuleTypeClass.EXPLORATION),
-    LOGICAL_JOIN_L_ASSCOM(RuleTypeClass.EXPLORATION),
-    LOGICAL_JOIN_L_ASSCOM_PROJECT(RuleTypeClass.EXPLORATION),
-    LOGICAL_JOIN_EXCHANGE(RuleTypeClass.EXPLORATION),
+    LOGICAL_INNER_JOIN_LASSCOM(RuleTypeClass.EXPLORATION),
+    LOGICAL_INNER_JOIN_LASSCOM_PROJECT(RuleTypeClass.EXPLORATION),
+    LOGICAL_OUTER_JOIN_LASSCOM(RuleTypeClass.EXPLORATION),
+    LOGICAL_OUTER_JOIN_LASSCOM_PROJECT(RuleTypeClass.EXPLORATION),
     LOGICAL_SEMI_JOIN_LOGICAL_JOIN_TRANSPOSE(RuleTypeClass.EXPLORATION),
     LOGICAL_SEMI_JOIN_LOGICAL_JOIN_TRANSPOSE_PROJECT(RuleTypeClass.EXPLORATION),
     LOGICAL_SEMI_JOIN_SEMI_JOIN_TRANPOSE(RuleTypeClass.EXPLORATION),
