+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/exploration/join/InnerJoinLAsscomProject.java
@@ -17,42 +17,15 @@
 
 package org.apache.doris.nereids.rules.exploration.join;
 
-import org.apache.doris.common.Pair;
 import org.apache.doris.nereids.rules.Rule;
 import org.apache.doris.nereids.rules.RuleType;
 import org.apache.doris.nereids.rules.exploration.OneExplorationRuleFactory;
-import org.apache.doris.nereids.rules.exploration.join.JoinReorderCommon.Type;
-import org.apache.doris.nereids.trees.plans.GroupPlan;
-import org.apache.doris.nereids.trees.plans.logical.LogicalJoin;
-import org.apache.doris.nereids.trees.plans.logical.LogicalProject;
-
-import java.util.function.Predicate;
 
 /**
  * Rule for change inner join LAsscom (associative and commutive).
  */
-public class JoinLAsscomProject extends OneExplorationRuleFactory {
-    // for inner-inner
-    public static final JoinLAsscomProject INNER = new JoinLAsscomProject(Type.INNER);
-    // for inner-leftOuter or leftOuter-leftOuter
-    public static final JoinLAsscomProject OUTER = new JoinLAsscomProject(Type.OUTER);
-
-    private final Predicate<LogicalJoin<LogicalProject<LogicalJoin<GroupPlan, GroupPlan>>, GroupPlan>> typeChecker;
-
-    private final Type type;
-
-    /**
-     * Specify join type.
-     */
-    public JoinLAsscomProject(Type type) {
-        this.type = type;
-        if (type == Type.INNER) {
-            typeChecker = join -> join.getJoinType().isInnerJoin() && join.left().child().getJoinType().isInnerJoin();
-        } else {
-            typeChecker = join -> JoinLAsscomHelper.outerSet.contains(
-                    Pair.of(join.left().child().getJoinType(), join.getJoinType()));
-        }
-    }
+public class InnerJoinLAsscomProject extends OneExplorationRuleFactory {
+    public static final InnerJoinLAsscomProject INSTANCE = new InnerJoinLAsscomProject();
 
     /*
      *        topJoin                   newTopJoin
@@ -65,9 +38,8 @@ public class JoinLAsscomProject extends OneExplorationRuleFactory {
      */
     @Override
     public Rule build() {
-        return logicalJoin(logicalProject(logicalJoin()), group())
-                .when(topJoin -> JoinLAsscomHelper.check(type, topJoin, topJoin.left().child()))
-                .when(typeChecker)
+        return innerLogicalJoin(logicalProject(innerLogicalJoin()), group())
+                .when(topJoin -> JoinLAsscomHelper.checkInner(topJoin, topJoin.left().child()))
                 .then(topJoin -> {
                     JoinLAsscomHelper helper = new JoinLAsscomHelper(topJoin, topJoin.left().child());
                     helper.initAllProject(topJoin.left());
@@ -75,6 +47,6 @@ public class JoinLAsscomProject extends OneExplorationRuleFactory {
                         return null;
                     }
                     return helper.newTopJoin();
-                }).toRule(RuleType.LOGICAL_JOIN_L_ASSCOM_PROJECT);
+                }).toRule(RuleType.LOGICAL_INNER_JOIN_LASSCOM_PROJECT);
     }
 }
