+++ fe/fe-core/src/test/java/org/apache/doris/nereids/rules/exploration/join/InnerJoinLAsscomProjectTest.java
@@ -21,6 +21,8 @@ import org.apache.doris.nereids.memo.Group;
 import org.apache.doris.nereids.memo.GroupExpression;
 import org.apache.doris.nereids.trees.expressions.EqualTo;
 import org.apache.doris.nereids.trees.expressions.Expression;
+import org.apache.doris.nereids.trees.expressions.NamedExpression;
+import org.apache.doris.nereids.trees.expressions.Slot;
 import org.apache.doris.nereids.trees.plans.JoinType;
 import org.apache.doris.nereids.trees.plans.Plan;
 import org.apache.doris.nereids.trees.plans.logical.LogicalJoin;
@@ -34,16 +36,18 @@ import com.google.common.collect.Lists;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
-public class JoinLAsscomTest {
+public class InnerJoinLAsscomProjectTest {
 
     private final LogicalOlapScan scan1 = PlanConstructor.newLogicalOlapScan(0, "t1", 0);
     private final LogicalOlapScan scan2 = PlanConstructor.newLogicalOlapScan(1, "t2", 0);
     private final LogicalOlapScan scan3 = PlanConstructor.newLogicalOlapScan(2, "t3", 0);
 
     @Test
-    public void testStarJoinLAsscom() {
+    public void testStarJoinLAsscomProject() {
         /*
          * Star-Join
          * t1 -- t2
@@ -53,8 +57,9 @@ public class JoinLAsscomTest {
          *     t1.id=t3.id               t1.id=t2.id
          *       topJoin                  newTopJoin
          *       /     \                   /     \
-         * t1.id=t2.id  t3          t1.id=t3.id   t2
-         * bottomJoin       -->    newBottomJoin
+         *    project   t3           project    project
+         * t1.id=t2.id             t1.id=t3.id    t2
+         * bottomJoin       -->   newBottomJoin
          *   /    \                   /    \
          * t1      t2               t1      t3
          */
@@ -65,12 +70,18 @@ public class JoinLAsscomTest {
         LogicalJoin<LogicalOlapScan, LogicalOlapScan> bottomJoin = new LogicalJoin<>(JoinType.INNER_JOIN,
                 Lists.newArrayList(bottomJoinOnCondition),
                 Optional.empty(), scan1, scan2);
-        LogicalJoin<LogicalJoin<LogicalOlapScan, LogicalOlapScan>, LogicalOlapScan> topJoin = new LogicalJoin<>(
-                JoinType.INNER_JOIN, Lists.newArrayList(topJoinOnCondition),
-                Optional.empty(), bottomJoin, scan3);
+
+        List<Slot> output = bottomJoin.getOutput();
+        List<NamedExpression> projectExprs = output.subList(0, output.size() - 1).stream()
+                .map(NamedExpression.class::cast).collect(Collectors.toList());
+        LogicalProject<LogicalJoin<LogicalOlapScan, LogicalOlapScan>> project = new LogicalProject<>(
+                projectExprs, bottomJoin);
+        LogicalJoin<LogicalProject<LogicalJoin<LogicalOlapScan, LogicalOlapScan>>, LogicalOlapScan>
+                topJoin = new LogicalJoin<>(JoinType.INNER_JOIN, Lists.newArrayList(topJoinOnCondition),
+                Optional.empty(), project, scan3);
 
         PlanChecker.from(MemoTestUtils.createConnectContext(), topJoin)
-                .transform(JoinLAsscom.INNER.build())
+                .transform(InnerJoinLAsscomProject.INSTANCE.build())
                 .checkMemo(memo -> {
                     Group root = memo.getRoot();
                     Assertions.assertEquals(2, root.getLogicalExpressions().size());
@@ -79,48 +90,21 @@ public class JoinLAsscomTest {
                     Assertions.assertTrue(root.logicalExpressionsAt(1).getPlan() instanceof LogicalProject);
 
                     GroupExpression newTopJoinGroupExpr = root.logicalExpressionsAt(1).child(0).getLogicalExpression();
-                    GroupExpression newBottomJoinGroupExpr = newTopJoinGroupExpr.child(0).getLogicalExpression();
-                    Plan bottomLeft = newBottomJoinGroupExpr.child(0).getLogicalExpression().getPlan();
-                    Plan bottomRight = newBottomJoinGroupExpr.child(1).getLogicalExpression().getPlan();
-                    Plan right = newTopJoinGroupExpr.child(1).getLogicalExpression().getPlan();
+                    GroupExpression leftProjectGroupExpr = newTopJoinGroupExpr.child(0).getLogicalExpression();
+                    GroupExpression rightProjectGroupExpr = newTopJoinGroupExpr.child(1).getLogicalExpression();
+                    Plan leftProject = newTopJoinGroupExpr.child(0).getLogicalExpression().getPlan();
+                    Plan rightProject = newTopJoinGroupExpr.child(1).getLogicalExpression().getPlan();
+                    Assertions.assertEquals(4, ((LogicalProject) leftProject).getProjects().size());
+                    Assertions.assertEquals(1, ((LogicalProject) rightProject).getProjects().size());
 
-                    Assertions.assertEquals("t1", ((LogicalOlapScan) bottomLeft).getTable().getName());
-                    Assertions.assertEquals("t3", ((LogicalOlapScan) bottomRight).getTable().getName());
-                    Assertions.assertEquals("t2", ((LogicalOlapScan) right).getTable().getName());
-                });
-    }
-
-    @Test
-    public void testChainJoinLAsscom() {
-        /*
-         * Chain-Join
-         * t1 -- t2 -- t3
-         * <p>
-         *     t2.id=t3.id               t2.id=t3.id
-         *       topJoin                  newTopJoin
-         *       /     \                   /     \
-         * t1.id=t2.id  t3          t1.id=t3.id   t2
-         * bottomJoin       -->    newBottomJoin
-         *   /    \                   /    \
-         * t1      t2               t1      t3
-         */
-
-        Expression bottomJoinOnCondition = new EqualTo(scan1.getOutput().get(0), scan2.getOutput().get(0));
-        Expression topJoinOnCondition = new EqualTo(scan2.getOutput().get(0), scan3.getOutput().get(0));
-        LogicalJoin<LogicalOlapScan, LogicalOlapScan> bottomJoin = new LogicalJoin<>(JoinType.INNER_JOIN,
-                Lists.newArrayList(bottomJoinOnCondition),
-                Optional.empty(), scan1, scan2);
-        LogicalJoin<LogicalJoin<LogicalOlapScan, LogicalOlapScan>, LogicalOlapScan> topJoin = new LogicalJoin<>(
-                JoinType.INNER_JOIN, Lists.newArrayList(topJoinOnCondition),
-                Optional.empty(), bottomJoin, scan3);
-
-        PlanChecker.from(MemoTestUtils.createConnectContext(), topJoin)
-                .transform(JoinLAsscom.INNER.build())
-                .checkMemo(memo -> {
-                    Group root = memo.getRoot();
-
-                    // TODO: need infer onCondition.
-                    Assertions.assertEquals(1, root.getLogicalExpressions().size());
+                    Plan t2 = rightProjectGroupExpr.child(0).getLogicalExpression().getPlan();
+                    Plan t1 = leftProjectGroupExpr.child(0).getLogicalExpression().child(0).getLogicalExpression()
+                            .getPlan();
+                    Plan t3 = leftProjectGroupExpr.child(0).getLogicalExpression().child(1).getLogicalExpression()
+                            .getPlan();
+                    Assertions.assertEquals("t2", ((LogicalOlapScan) t2).getTable().getName());
+                    Assertions.assertEquals("t1", ((LogicalOlapScan) t1).getTable().getName());
+                    Assertions.assertEquals("t3", ((LogicalOlapScan) t3).getTable().getName());
                 });
     }
 }
